+++ app/forms/already_filed_form.rb
@@ -1,8 +1,7 @@
 class AlreadyFiledForm < QuestionsForm
-  set_attributes_for :intake, :already_filed
+  set_attributes_for :intake, :already_filed, :source, :referrer
 
   def save
     @intake.update(attributes_for(:intake))
   end
-
 end
