+++ spec/features/web_intake/requested_documents_token_spec.rb
@@ -43,6 +43,6 @@ RSpec.feature "Client uploads a requested document" do
     expect(page).to have_button("Continue", disabled: true)
 
     visit "/questions/job-count"
-    expect(current_path).to eq("/questions/feelings")
+    expect(current_path).to eq("/questions/welcome")
   end
 end
